bottle: add docs & import protection
